+++ src/main/java/view/RetryOnException.java
@@ -1,14 +1,14 @@
-package utils;
+package view;
 
 import java.util.function.Supplier;
 
-public class Retry {
-    public static <T> T retryOnException(Supplier<T> supplier) {
+public class RetryOnException {
+    public static <T> T retryInputOnIllegalArgumentException(Supplier<T> supplier) {
         try {
             return supplier.get();
         } catch (IllegalArgumentException e) {
             System.out.println("[ERROR] " + e.getMessage());
-            return retryOnException(supplier);
+            return retryInputOnIllegalArgumentException(supplier);
         }
     }
 }
